Backend Java/s3-individualassignment-main/s3-individualassignment-main/src/main/java/com/example/demo/business/Converters/TicketConverter.java
package com.example.demo.business.Converters;

import com.example.demo.domain.Tickets;
import com.example.demo.repository.TicketEntity;

public class TicketConverter {
    private TicketConverter(){

    }

    public static Tickets convert(TicketEntity ticket){
        return Tickets.builder()
                .id(ticket.getId())
                .match(MatchConverter.convert(ticket.getMatch()))
                .date(ticket.getDate())
                .place(ticket.getPlace())
                .price(ticket.getPrice())
                .quantity(ticket.getQuantity())
                .build();
    }
}
